# docs/swagger.yaml
basePath: /
definitions:
  gin-scalable-api_pkg_response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_modules_company.CompanyListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_company.CompanyResponse'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_modules_subscription.AddModulesToPlanRequest:
    properties:
      module_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - module_ids
    type: object
  internal_modules_subscription.CreateSubscriptionPlanRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      features:
        additionalProperties: true
        type: object
      max_branches:
        type: integer
      max_users:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      price_monthly:
        minimum: 0
        type: number
      price_yearly:
        minimum: 0
        type: number
    required:
    - display_name
    - name
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Complete API for ERP with RBAC and API Documentation System. This API provides comprehensive endpoints for managing users, companies, branches, roles, modules, subscriptions, units, applications, and authentication with JWT-based security.

    ## Environments
    - **Development**: http://localhost:8081
    - **Staging**: https://staging-api.huminor.com
    - **Production**: https://api.huminor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Huminor Console API
  version: "1.0"
paths:
...
  /api/v1/admin/plan-modules/{plan_id}:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar modules yang termasuk dalam subscription plan
        (admin only)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: plan_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Plan modules berhasil diambil
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid plan ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get plan modules (Admin)
      tags:
      - Subscription Plans (Admin)
    post:
      consumes:
      - application/json
      description: Menambahkan modules ke subscription plan (admin only)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: plan_id
        required: true
        type: integer
      - description: Modules to add
        in: body
        name: modules
        required: true
        schema:
          $ref: '#/definitions/internal_modules_subscription.AddModulesToPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Modules berhasil ditambahkan ke plan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid plan ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Subscription plan tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Add modules to plan (Admin)
      tags:
      - Subscription Plans (Admin)
  /api/v1/admin/plan-modules/{plan_id}/{module_id}:
    delete:
      consumes:
      - application/json
      description: Menghapus module dari subscription plan (admin only)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: plan_id
        required: true
        type: integer
      - description: Module ID
        in: path
        name: module_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Module berhasil dihapus dari plan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Plan atau module tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove module from plan (Admin)
      tags:
      - Subscription Plans (Admin)
  /api/v1/admin/subscription-plans:
    post:
      consumes:
      - application/json
      description: Membuat subscription plan baru (admin only)
      parameters:
      - description: Subscription plan data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/internal_modules_subscription.CreateSubscriptionPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Subscription plan berhasil dibuat
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_subscription.SubscriptionPlanResponse'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "409":
          description: Conflict - plan name sudah ada
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create subscription plan (Admin)
      tags:
      - Subscription Plans (Admin)
  /api/v1/admin/subscription-plans/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus subscription plan (admin only)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subscription plan berhasil dihapus
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid plan ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Subscription plan tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete subscription plan (Admin)
      tags:
      - Subscription Plans (Admin)
    put:
      consumes:
      - application/json
      description: Memperbarui subscription plan (admin only)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription plan data yang akan diupdate
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/internal_modules_subscription.UpdateSubscriptionPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription plan berhasil diupdate
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_subscription.SubscriptionPlanResponse'
              type: object
        "400":
          description: Bad request - Invalid plan ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Subscription plan tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update subscription plan (Admin)
      tags:
      - Subscription Plans (Admin)
...
  /api/v1/companies:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua company dengan filter opsional dan pagination
      parameters:
      - description: Limit jumlah data
        in: query
        name: limit
        type: integer
      - description: Offset data
        in: query
        name: offset
        type: integer
      - description: Search by name atau code
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Companies berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get all companies
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Membuat company baru
      parameters:
      - description: Company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/internal_modules_company.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Company berhasil dibuat
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyResponse'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "409":
          description: Conflict - company code sudah ada
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create new company
      tags:
      - Companies
...
  /api/v1/companies/{id}/subscription:
    get:
      consumes:
      - application/json
      description: Mendapatkan subscription aktif dari company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Company subscription berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_subscription.SubscriptionResponse'
              type: object
        "400":
          description: Bad request - Invalid company ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Subscription tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get company subscription
      tags:
      - Subscriptions
...
  /api/v1/subscription-plans:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua subscription plans (public endpoint)
      produces:
      - application/json
      responses:
        "200":
          description: Subscription plans berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_subscription.SubscriptionPlanResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      summary: Get all subscription plans
      tags:
      - Subscription Plans
  /api/v1/subscription-plans/{id}:
    get:
      consumes:
      - application/json
      description: Mendapatkan detail subscription plan berdasarkan ID (public endpoint)
      parameters:
      - description: Subscription Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subscription plan berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_subscription.SubscriptionPlanResponse'
              type: object
        "400":
          description: Bad request - Invalid plan ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Subscription plan tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      summary: Get subscription plan by ID
      tags:
      - Subscription Plans
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...